Move Preferences and Profile controllers under Users\
