+++ translations/errors+intl-icu.en_GB.yaml
@@ -62,6 +62,5 @@ user.email.invalid: 'Enter a valid email address.'
 user.email.required: 'Enter an email address.'
 user.language.invalid: 'Select a language from the list.'
 user.language.required: 'Select a language.'
-user.ldap.cannot_update_profile: 'You can’t update your profile because it’s managed by LDAP.'
 user.name.max_chars: 'Enter a name of less than {limit} characters.'
 user.password.dont_match: 'The password does not match, please try with a different one.'
