+++ translations/errors+intl-icu.fr_FR.yaml
@@ -62,6 +62,5 @@ user.email.invalid: 'Saisissez une adresse email valide.'
 user.email.required: 'Saisissez une adresse email.'
 user.language.invalid: 'Sélectionnez une langue de la liste.'
 user.language.required: 'Sélectionnez une langue.'
-user.ldap.cannot_update_profile: 'Vous ne pouvez pas modifier votre profil car il est géré avec LDAP.'
 user.name.max_chars: 'Saisissez un nom de moins de {limit} caractères.'
 user.password.dont_match: 'Le mot de passe ne correspond pas, veuillez essayer avec un différent.'
